+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [14.2.5-dev.1] - 2019-08-21]
+
 - [#1912](https://github.com/teambit/bit/issues/1912) fix linking to an internal and non-exist dist file to fallback to the main package
 - [#1937](https://github.com/teambit/bit/issues/1937) introduce `--json` flag for `bit import`
 - [#1916](https://github.com/teambit/bit/issues/1916) do not throw an error when there are no components to build